🐛 fix rating bug and psychologist pagination

+++ app/src/main/java/com/lans/sleep_care/data/source/network/dto/request/therapy/RatingTherapyRequest.kt
@@ -5,6 +5,6 @@ import com.squareup.moshi.Json
 data class RatingTherapyRequest(
     @field:Json(name = "therapy_id")
     val therapyId: Int,
-    val rating: Int,
+    val rating: Double,
     val comment: String
 )
