+++ app/src/main/java/com/lans/sleep_care/domain/usecase/therapy/CreateRatingUseCase.kt
@@ -4,5 +4,5 @@ import com.lans.sleep_care.data.Resource
 import kotlinx.coroutines.flow.Flow
 
 interface CreateRatingUseCase {
-    suspend fun execute(therapyId: Int, rating: Int, comment: String): Flow<Resource<Boolean>>
+    suspend fun execute(therapyId: Int, rating: Double, comment: String): Flow<Resource<Boolean>>
 }
