+++ app/src/main/java/com/lans/sleep_care/presentation/screen/history_detail/HistoryDetailUIState.kt
@@ -5,7 +5,7 @@ import com.lans.sleep_care.domain.model.therapy.Therapy
 import com.lans.sleep_care.domain.model.therapy.TherapySchedule
 
 data class HistoryDetailUIState(
-    var rating: Int = 0,
+    var rating: Double = 0.0,
     var comment: String = "",
     var isLoading: Boolean = false,
     var error: String = "",
